% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{class_packages}
\alias{class_packages}
\title{Required packages for SDS courses}
\format{
A named list where:
\describe{
\item{Names}{SDS course codes (e.g. \"sds_100\")}
\item{Values}{Character vectors of package dependencies}
}
}
\source{
Manually curated by package maintainers
}
\usage{
data(class_packages)
}
\description{
A named list mapping SDS course codes to required R packages.
}
\keyword{datasets}
