% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdsload-package.R
\docType{package}
\name{sdsload-package}
\alias{sdsload}
\alias{sdsload-package}
\title{sdsload: A package that loads all necessary packages for a selection of Smith College Statistical and Data Science classes}
\description{
A package that loads all necessary packages for a selection of Smith College Statistical and Data Science classes. Current offerings include SDS100, SDS192, SDS220, and SDS 291.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gabriellekraft/sdsload}
  \item Report bugs at \url{https://github.com/gabriellekraft/sdsload/issues}
}

}
\author{
\strong{Maintainer}: Gabrielle Kraft \email{gkraft@smith.edu}

Authors:
\itemize{
  \item Sofija sstankovic@smith.edu Stankovic
}

}
\keyword{internal}
