% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{class_packages}
\alias{class_packages}
\title{Core packages for selected Smith College SDS courses}
\source{
Manually curated by package maintainers
}
\usage{
data(class_packages)
}
\description{
A named list mapping SDS course codes to core R packages.
}
\keyword{internal}
