% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdsload.R
\name{load_sds}
\alias{load_sds}
\title{Do the annotations here bae}
\usage{
load_sds(sds_class, install_missing = TRUE)
}
\description{
Given a comic number, the \code{xkcd()} function calls the xkcd JSON API and returns metadata about the comic in the form of a list object.
}
